Updated Game Dashboard to Figma Design
