+++ web/src/components/CCSS.tsx
@@ -3,10 +3,10 @@ import Typography from '@material-ui/core/Typography';
 import { Game } from '../API';
 
 export default function CCSS(game: Game) {
-  if (game.grade === 'General') return <Typography color="textSecondary" gutterBottom>{game.grade}</Typography>;
+  if (game.grade === 'General') return <Typography style={{fontWeight: 700, color: '#9139F8'}} >{game.grade}</Typography>;
   if (game.grade && game.domain && game.cluster && game.standard) {
     return (
-      <Typography color="textSecondary" gutterBottom>
+      <Typography style={{fontWeight: 700, color: '#9139F8'}}>
         {`${game.grade}.${game.domain}.${game.cluster}.${game.standard}`}
       </Typography>
     );
