+++ web/src/lib/questions.ts
@@ -1,5 +1,6 @@
 import { API, graphqlOperation } from 'aws-amplify';
 import { updateQuestion as UQ, createQuestion as CQ, createGameQuestion as CGQ } from '../graphql/mutations';
+import { listQuestions } from '../graphql/queries';
 
 export const createQuestion = async(question: any, gameId: any) => {
   const createdQuestion = await API.graphql(graphqlOperation(CQ, { question })) as { data: any };
@@ -13,4 +14,10 @@ export const updateQuestion = async (question: any) => {
   delete question.createdAt;
   const result = await API.graphql(graphqlOperation(UQ, { question })) as { data: any };
   return result?.data?.updateQuestion || [];
+}
+
+export const listOfQuestions = async () => {
+  const listedQuestions = await API.graphql(graphqlOperation(listQuestions)) as { data: any };
+  // console.log(listedQuestions);
+  return listedQuestions.data.listQuestions;
 };
